# docs/api/xorb-openapi-spec.yaml
openapi: 3.1.0
info:
  title: XORB Cybersecurity Platform API
  description: |
    XORB API for autonomous cybersecurity operations, orchestration, and multi-agent coordination.
    
    ## Security
    - All endpoints require mTLS authentication
    - JWT bearer tokens for authorization
    - Role-based access control (RBAC)
    - Request signing for response integrity
    
    ## Rate Limiting
    - Admin: 10,000 req/min
    - Orchestrator: 5,000 req/min  
    - Agent: 2,000 req/min
    - Analyst: 1,000 req/min
    - Readonly: 500 req/min
    
  version: 1.0.0
  contact:
    name: XORB Platform Team
    email: platform@xorb.security
  license:
    name: Proprietary
servers:
  - url: https://api.xorb.security/v1
    description: Production API
  - url: https://staging-api.xorb.security/v1
    description: Staging API
  - url: http://localhost:8000/v1
    description: Development API
security:
  - BearerAuth: []
  - mTLS: []
paths:
...
  /auth/token:
    post:
      summary: Obtain JWT token
      description: Exchange client credentials for JWT token
      tags: [Authentication]
      security: 
        - mTLS: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mTLS:
      type: mutualTLS
      description: Client certificate authentication
  schemas:
...
    TokenRequest:
      type: object
      required: [client_id, client_secret]
      properties:
        client_id:
          type: string
          description: Client identifier
        client_secret:
          type: string
          description: Client secret
        scope:
          type: string
          description: Requested permissions scope
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token expiry in seconds
        scope:
          type: string
          description: Granted permissions

    # Agent Management Schemas
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...